BridgeJS: Split out OptionalSupportTests.swift
