+++ Tests/BridgeJSRuntimeTests/JavaScript/Types.mjs
@@ -1,0 +1,6 @@
+export class ImportedFoo {
+    /** @param {string} value */
+    constructor(value) {
+        this.value = value;
+    }
+}
